refactored application type files and reorganized file structure
