+++ src/main/calendar/calendarMonth.css
@@ -1,4 +1,4 @@
-.calendar {
+.calendarMonth {
     border: 1px solid black;
 }
