+++ src/main/calendar/calendarTypes.tsx
@@ -1,0 +1,5 @@
+export interface StockValue {
+    symbol: string,
+    value: number,
+    type: 'actual' | 'est'
+};
