+++ src/main/dividend search/dividendSearch.css
@@ -1,0 +1,20 @@
+#dividendSearch form {
+    width: 650px;
+    margin: 50px auto 30px;
+}
+
+#dividendSearch input {
+    width: 45px;
+    margin-right: 20px;
+}
+
+    #dividendSearch input[type="submit"] {
+        width: 49px;
+    }
+
+.trackedStockPosition-container {
+    display: flex;
+    flex-wrap: wrap;
+    width: 650px;
+    margin: 0px auto 60px;
+}
