+++ src/main/dividend search/trackedStockPosition.css
@@ -1,25 +1,3 @@
-#dividendSearch form {
-	width: 650px;
-	margin: 50px auto 30px;
-}
-
-#dividendSearch input {
-	width: 45px;
-	margin-right: 20px;
-}
-
-#dividendSearch input[type="submit"]{
-	width: 49px;
-}
-
-.trackedStockPosition-container{
-	display: flex;
-	flex-wrap: wrap;
-	width: 650px;
-	margin: 0px auto 60px;
-
-}
-
 .trackedStockPosition {
 	flex-basis: 130px;
 	margin-top: 20px;
