--- src/testing/testData.tsx
@@ -1,648 +1,0 @@
-import { DividendRequestData, StockPosition } from '../interfaces';
-
-export const testStockPositions: StockPosition[] = [
-    { symbol: 'T', shares: 1179 },{ symbol: 'MO', shares:  754},{ symbol: 'ED', shares: 580 },{ symbol: 'SO', shares:  687},{ symbol: 'ETR', shares:  415},
-    { symbol: 'D', shares:  569},{ symbol: 'PPL', shares:  1535},{ symbol: 'DUK', shares:  447},{ symbol: 'KMI', shares: 2065 },{ symbol: 'XOM', shares: 630 }
-]
-
-/*This response data format is outdated
-export const testDividendData: DividendRequestData = {
-    "dividendCalendarList": [
-        {
-            "type": "actual",
-            "symbol": "PPL",
-            "shares": 1535,
-            "frequency": "quarterly",
-            "amountPerShare": 0.415,
-            "exDate": "2020-12-09",
-            "paymentDate": "2021-01-04",
-            "paymentYear": 2021,
-            "paymentMonth": 1,
-            "amountTotal": 637.025
-        },
-        {
-            "type": "actual",
-            "symbol": "MO",
-            "shares": 754,
-            "frequency": "quarterly",
-            "amountPerShare": 0.86,
-            "exDate": "2020-12-24",
-            "paymentDate": "2021-01-11",
-            "paymentYear": 2021,
-            "paymentMonth": 1,
-            "amountTotal": 648.4399999999999
-        },
-        {
-            "type": "actual",
-            "symbol": "T",
-            "shares": 1179,
-            "frequency": "quarterly",
-            "amountPerShare": 0.52,
-            "exDate": "2021-01-08",
-            "paymentDate": "2021-02-01",
-            "paymentYear": 2021,
-            "paymentMonth": 2,
-            "amountTotal": 613.08
-        },
-        {
-            "type": "actual",
-            "symbol": "KMI",
-            "shares": 2065,
-            "frequency": "quarterly",
-            "amountPerShare": 0.2625,
-            "exDate": "2021-01-29",
-            "paymentDate": "2021-02-16",
-            "paymentYear": 2021,
-            "paymentMonth": 2,
-            "amountTotal": 542.0625
-        },
-        {
-            "type": "actual",
-            "symbol": "ETR",
-            "shares": 415,
-            "frequency": "quarterly",
-            "amountPerShare": 0.95,
-            "exDate": "2021-02-11",
-            "paymentDate": "2021-03-01",
-            "paymentYear": 2021,
-            "paymentMonth": 3,
-            "amountTotal": 394.25
-        },
-        {
-            "type": "actual",
-            "symbol": "SO",
-            "shares": 687,
-            "frequency": "quarterly",
-            "amountPerShare": 0.64,
-            "exDate": "2021-02-12",
-            "paymentDate": "2021-03-08",
-            "paymentYear": 2021,
-            "paymentMonth": 3,
-            "amountTotal": 439.68
-        },
-        {
-            "type": "actual",
-            "symbol": "XOM",
-            "shares": 630,
-            "frequency": "quarterly",
-            "amountPerShare": 0.87,
-            "exDate": "2021-02-09",
-            "paymentDate": "2021-03-10",
-            "paymentYear": 2021,
-            "paymentMonth": 3,
-            "amountTotal": 548.1
-        },
-        {
-            "type": "actual",
-            "symbol": "ED",
-            "shares": 580,
-            "frequency": "quarterly",
-            "amountPerShare": 0.775,
-            "exDate": "2021-02-16",
-            "paymentDate": "2021-03-15",
-            "paymentYear": 2021,
-            "paymentMonth": 3,
-            "amountTotal": 449.5
-        },
-        {
-            "type": "actual",
-            "symbol": "DUK",
-            "shares": 447,
-            "frequency": "quarterly",
-            "amountPerShare": 0.965,
-            "exDate": "2021-02-11",
-            "paymentDate": "2021-03-16",
-            "paymentYear": 2021,
-            "paymentMonth": 3,
-            "amountTotal": 431.35499999999996
-        },
-        {
-            "type": "actual",
-            "symbol": "D",
-            "shares": 569,
-            "frequency": "quarterly",
-            "amountPerShare": 0.63,
-            "exDate": "2021-03-04",
-            "paymentDate": "2021-03-20",
-            "paymentYear": 2021,
-            "paymentMonth": 3,
-            "amountTotal": 358.47
-        },
-        {
-            "type": "actual",
-            "symbol": "PPL",
-            "shares": 1535,
-            "frequency": "quarterly",
-            "amountPerShare": 0.415,
-            "exDate": "2021-03-09",
-            "paymentDate": "2021-04-01",
-            "paymentYear": 2021,
-            "paymentMonth": 4,
-            "amountTotal": 637.025
-        },
-        {
-            "type": "actual",
-            "symbol": "MO",
-            "shares": 754,
-            "frequency": "quarterly",
-            "amountPerShare": 0.86,
-            "exDate": "2021-03-24",
-            "paymentDate": "2021-04-30",
-            "paymentYear": 2021,
-            "paymentMonth": 4,
-            "amountTotal": 648.4399999999999
-        },
-        {
-            "type": "actual",
-            "symbol": "T",
-            "shares": 1179,
-            "frequency": "quarterly",
-            "amountPerShare": 0.52,
-            "exDate": "2021-04-08",
-            "paymentDate": "2021-05-03",
-            "paymentYear": 2021,
-            "paymentMonth": 5,
-            "amountTotal": 613.08
-        },
-        {
-            "type": "actual",
-            "symbol": "KMI",
-            "shares": 2065,
-            "frequency": "quarterly",
-            "amountPerShare": 0.27,
-            "exDate": "2021-04-29",
-            "paymentDate": "2021-05-17",
-            "paymentYear": 2021,
-            "paymentMonth": 5,
-            "amountTotal": 557.5500000000001
-        },
-        {
-            "type": "actual",
-            "symbol": "ETR",
-            "shares": 415,
-            "frequency": "quarterly",
-            "amountPerShare": 0.95,
-            "exDate": "2021-05-05",
-            "paymentDate": "2021-06-01",
-            "paymentYear": 2021,
-            "paymentMonth": 6,
-            "amountTotal": 394.25
-        },
-        {
-            "type": "actual",
-            "symbol": "SO",
-            "shares": 687,
-            "frequency": "quarterly",
-            "amountPerShare": 0.66,
-            "exDate": "2021-05-14",
-            "paymentDate": "2021-06-07",
-            "paymentYear": 2021,
-            "paymentMonth": 6,
-            "amountTotal": 453.42
-        },
-        {
-            "type": "actual",
-            "symbol": "XOM",
-            "shares": 630,
-            "frequency": "quarterly",
-            "amountPerShare": 0.87,
-            "exDate": "2021-05-12",
-            "paymentDate": "2021-06-10",
-            "paymentYear": 2021,
-            "paymentMonth": 6,
-            "amountTotal": 548.1
-        },
-        {
-            "type": "actual",
-            "symbol": "ED",
-            "shares": 580,
-            "frequency": "quarterly",
-            "amountPerShare": 0.775,
-            "exDate": "2021-05-18",
-            "paymentDate": "2021-06-15",
-            "paymentYear": 2021,
-            "paymentMonth": 6,
-            "amountTotal": 449.5
-        },
-        {
-            "type": "actual",
-            "symbol": "DUK",
-            "shares": 447,
-            "frequency": "quarterly",
-            "amountPerShare": 0.965,
-            "exDate": "2021-05-13",
-            "paymentDate": "2021-06-16",
-            "paymentYear": 2021,
-            "paymentMonth": 6,
-            "amountTotal": 431.35499999999996
-        },
-        {
-            "type": "actual",
-            "symbol": "D",
-            "shares": 569,
-            "frequency": "quarterly",
-            "amountPerShare": 0.63,
-            "exDate": "2021-06-03",
-            "paymentDate": "2021-06-20",
-            "paymentYear": 2021,
-            "paymentMonth": 6,
-            "amountTotal": 358.47
-        },
-        {
-            "type": "actual",
-            "symbol": "PPL",
-            "shares": 1535,
-            "frequency": "quarterly",
-            "amountPerShare": 0.415,
-            "exDate": "2021-06-09",
-            "paymentDate": "2021-07-01",
-            "paymentYear": 2021,
-            "paymentMonth": 7,
-            "amountTotal": 637.025
-        },
-        {
-            "type": "actual",
-            "symbol": "MO",
-            "shares": 754,
-            "frequency": "quarterly",
-            "amountPerShare": 0.86,
-            "exDate": "2021-06-14",
-            "paymentDate": "2021-07-09",
-            "paymentYear": 2021,
-            "paymentMonth": 7,
-            "amountTotal": 648.4399999999999
-        },
-        {
-            "type": "actual",
-            "symbol": "T",
-            "shares": 1179,
-            "frequency": "quarterly",
-            "amountPerShare": 0.52,
-            "exDate": "2021-07-08",
-            "paymentDate": "2021-08-02",
-            "paymentYear": 2021,
-            "paymentMonth": 8,
-            "amountTotal": 613.08
-        },
-        {
-            "type": "est",
-            "symbol": "KMI",
-            "shares": 2065,
-            "frequency": "quarterly",
-            "amountPerShare": 0.27,
-            "exDate": "2021-07-29",
-            "paymentDate": "2021-08-17",
-            "paymentYear": 2021,
-            "paymentMonth": 8,
-            "amountTotal": 557.5500000000001
-        },
-        {
-            "type": "est",
-            "symbol": "ETR",
-            "shares": 415,
-            "frequency": "quarterly",
-            "amountPerShare": 0.95,
-            "exDate": "2021-08-05",
-            "paymentDate": "2021-09-01",
-            "paymentYear": 2021,
-            "paymentMonth": 9,
-            "amountTotal": 394.25
-        },
-        {
-            "type": "est",
-            "symbol": "SO",
-            "shares": 687,
-            "frequency": "quarterly",
-            "amountPerShare": 0.66,
-            "exDate": "2021-08-14",
-            "paymentDate": "2021-09-07",
-            "paymentYear": 2021,
-            "paymentMonth": 9,
-            "amountTotal": 453.42
-        },
-        {
-            "type": "est",
-            "symbol": "XOM",
-            "shares": 630,
-            "frequency": "quarterly",
-            "amountPerShare": 0.87,
-            "exDate": "2021-08-12",
-            "paymentDate": "2021-09-10",
-            "paymentYear": 2021,
-            "paymentMonth": 9,
-            "amountTotal": 548.1
-        },
-        {
-            "type": "est",
-            "symbol": "ED",
-            "shares": 580,
-            "frequency": "quarterly",
-            "amountPerShare": 0.775,
-            "exDate": "2021-08-18",
-            "paymentDate": "2021-09-15",
-            "paymentYear": 2021,
-            "paymentMonth": 9,
-            "amountTotal": 449.5
-        },
-        {
-            "type": "est",
-            "symbol": "DUK",
-            "shares": 447,
-            "frequency": "quarterly",
-            "amountPerShare": 0.965,
-            "exDate": "2021-08-13",
-            "paymentDate": "2021-09-16",
-            "paymentYear": 2021,
-            "paymentMonth": 9,
-            "amountTotal": 431.35499999999996
-        },
-        {
-            "type": "est",
-            "symbol": "D",
-            "shares": 569,
-            "frequency": "quarterly",
-            "amountPerShare": 0.63,
-            "exDate": "2021-09-03",
-            "paymentDate": "2021-09-20",
-            "paymentYear": 2021,
-            "paymentMonth": 9,
-            "amountTotal": 358.47
-        },
-        {
-            "type": "est",
-            "symbol": "PPL",
-            "shares": 1535,
-            "frequency": "quarterly",
-            "amountPerShare": 0.415,
-            "exDate": "2021-09-09",
-            "paymentDate": "2021-10-01",
-            "paymentYear": 2021,
-            "paymentMonth": 10,
-            "amountTotal": 637.025
-        },
-        {
-            "type": "est",
-            "symbol": "MO",
-            "shares": 754,
-            "frequency": "quarterly",
-            "amountPerShare": 0.86,
-            "exDate": "2021-09-14",
-            "paymentDate": "2021-10-09",
-            "paymentYear": 2021,
-            "paymentMonth": 10,
-            "amountTotal": 648.4399999999999
-        },
-        {
-            "type": "est",
-            "symbol": "T",
-            "shares": 1179,
-            "frequency": "quarterly",
-            "amountPerShare": 0.52,
-            "exDate": "2021-10-08",
-            "paymentDate": "2021-11-02",
-            "paymentYear": 2021,
-            "paymentMonth": 11,
-            "amountTotal": 613.08
-        },
-        {
-            "type": "est",
-            "symbol": "KMI",
-            "shares": 2065,
-            "frequency": "quarterly",
-            "amountPerShare": 0.27,
-            "exDate": "2021-10-29",
-            "paymentDate": "2021-11-17",
-            "paymentYear": 2021,
-            "paymentMonth": 11,
-            "amountTotal": 557.5500000000001
-        },
-        {
-            "type": "est",
-            "symbol": "ETR",
-            "shares": 415,
-            "frequency": "quarterly",
-            "amountPerShare": 0.95,
-            "exDate": "2021-11-05",
-            "paymentDate": "2021-12-01",
-            "paymentYear": 2021,
-            "paymentMonth": 12,
-            "amountTotal": 394.25
-        },
-        {
-            "type": "est",
-            "symbol": "SO",
-            "shares": 687,
-            "frequency": "quarterly",
-            "amountPerShare": 0.66,
-            "exDate": "2021-11-14",
-            "paymentDate": "2021-12-07",
-            "paymentYear": 2021,
-            "paymentMonth": 12,
-            "amountTotal": 453.42
-        },
-        {
-            "type": "est",
-            "symbol": "XOM",
-            "shares": 630,
-            "frequency": "quarterly",
-            "amountPerShare": 0.87,
-            "exDate": "2021-11-12",
-            "paymentDate": "2021-12-10",
-            "paymentYear": 2021,
-            "paymentMonth": 12,
-            "amountTotal": 548.1
-        },
-        {
-            "type": "est",
-            "symbol": "ED",
-            "shares": 580,
-            "frequency": "quarterly",
-            "amountPerShare": 0.775,
-            "exDate": "2021-11-18",
-            "paymentDate": "2021-12-15",
-            "paymentYear": 2021,
-            "paymentMonth": 12,
-            "amountTotal": 449.5
-        },
-        {
-            "type": "est",
-            "symbol": "DUK",
-            "shares": 447,
-            "frequency": "quarterly",
-            "amountPerShare": 0.965,
-            "exDate": "2021-11-13",
-            "paymentDate": "2021-12-16",
-            "paymentYear": 2021,
-            "paymentMonth": 12,
-            "amountTotal": 431.35499999999996
-        },
-        {
-            "type": "est",
-            "symbol": "D",
-            "shares": 569,
-            "frequency": "quarterly",
-            "amountPerShare": 0.63,
-            "exDate": "2021-12-03",
-            "paymentDate": "2021-12-20",
-            "paymentYear": 2021,
-            "paymentMonth": 12,
-            "amountTotal": 358.47
-        }
-    ],
-    "dividendCalendarAgList": [
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "PPL",
-                "MO"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 1,
-            "amountTotal": 1285.465
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "T",
-                "KMI"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 2,
-            "amountTotal": 1155.1425
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "ETR",
-                "SO",
-                "XOM",
-                "ED",
-                "DUK",
-                "D"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 3,
-            "amountTotal": 2621.3550000000005
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "PPL",
-                "MO"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 4,
-            "amountTotal": 1285.465
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "T",
-                "KMI"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 5,
-            "amountTotal": 1170.63
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "ETR",
-                "SO",
-                "XOM",
-                "ED",
-                "DUK",
-                "D"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 6,
-            "amountTotal": 2635.0950000000003
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "PPL",
-                "MO"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 7,
-            "amountTotal": 1285.465
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "T",
-                "KMI"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 8,
-            "amountTotal": 1170.63
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "ETR",
-                "SO",
-                "XOM",
-                "ED",
-                "DUK",
-                "D"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 9,
-            "amountTotal": 2635.0950000000003
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "PPL",
-                "MO"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 10,
-            "amountTotal": 1285.465
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "T",
-                "KMI"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 11,
-            "amountTotal": 1170.63
-        },
-        {
-            "agType": "subtotal",
-            "symbolList": [
-                "ETR",
-                "SO",
-                "XOM",
-                "ED",
-                "DUK",
-                "D"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": 12,
-            "amountTotal": 2635.0950000000003
-        },
-        {
-            "agType": "total",
-            "symbolList": [
-                "T",
-                "MO",
-                "ED",
-                "SO",
-                "ETR",
-                "D",
-                "PPL",
-                "DUK",
-                "KMI",
-                "XOM"
-            ],
-            "paymentYear": 2021,
-            "paymentMonth": -1,
-            "amountTotal": 20335.532499999994
-        }
-    ]
-};*/
